refactor: 공고 중복파일 제거
